Working point toolbar
Point toolbar successfully adds point to the points array inside the
country of choice inside the countries array in the map. New fix gives a
constant eventListener to the map so that whenever the user clicks the
map, the point toolbar coordinates update. From the point toolbar, the
user can edit the coordinates, choose a country to add the new point to,
and then add it.